OrthoViewExample.kt, OrhtoView: Move Ortho View Method to own file
